Create MostSequenceWord.java
Given a text. Count n words with the most frequent occurrences (case insensitive, i.e Word, WORD or WoRd are the same). If there are many words with the same frequency, select the words with the smallest alphabetical order (for example, “an” has the lower  alphabetical order than “bow”).

A word is defined as a set of consecutive letters. All non-letter characters are consider as delimiters.

Input/output:

Input:

Text with no more than 50000 characters and number n (1 <= n <= 1000). All characters in the text are ASCII characters

Output:

List of n words with the most frequent occurrences, which is arranged in descending order of frequency, if the words have equal frequency, the words with the lower alphabetical order will be preceded (each word in the list only including lowercase letters. If there are not enough n words in the input text, add empty strings to the end of the list.